Move user_cache_dir from constants to config
