+++ src/atldld/cli.py
@@ -59,7 +59,7 @@ def cache_dir():
     """Print the location of the global cache directory."""
     import os
 
-    from atldld.constants import user_cache_dir
+    from atldld.config import user_cache_dir
 
     if "XDG_CACHE_HOME" in os.environ:
         suffix = " (configured via XDG_CACHE_HOME)"
